web_builder/requirements.txt
flask==2.2.3
flask-jwt-extended==4.4.4
python-dotenv==1.0.0
pymongo==4.3.3
requests==2.28.2
gunicorn==20.1.0
python-dateutil==2.8.2
bcrypt==4.0.1
bson==0.5.10
dnspython==2.3.0